meet-live-transcriber/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=0.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
